backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
minio==7.2.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
onnxruntime==1.17.0
PyJWT==2.12.1
cryptography==41.0.7
email-validator==2.1.0
